content-moderation-openenv/requirements.txt
fastapi
uvicorn
pydantic
openai
python-dotenv
requests
pytest

